Created proc macro for linker_script
In this change I'm exploring the idea of automatically generating the
layout through a definition given by a proc-macro. Highlighting errors
would be easy and nice for the user to make sense of the problem.

On the other hand, having all the information about the map statically
defined (even though some identifiers might still be computed at
runtime) allows to do some more extensive validation of the linker
script before actually generating it.

This is still WIP, just saving the progress for now.